+++ test/blackbox-tests/test-cases/coq/compose-boot.t/B/b.v
@@ -1,0 +1,1 @@
+Inductive Begining := Of | The | Universe.