+++ test/blackbox-tests/test-cases/coq/compose-installed.t/run.t
@@ -1,0 +1,20 @@
+We test composing a project with an installed Coq theory. The installed theory
+does *not* have to be a dune project.
+
+  $ export COQLIB=.
+
+TODO: Currently this is not supported so the output is garbage
+
+  $ cd A
+  $ dune build
+  File "dune", line 3, characters 11-12:
+  3 |  (theories B))
+                 ^
+  Theory B not found
+  -> required by theory A in .
+  -> required by _build/default/a.v.d
+  -> required by _build/default/.a.aux
+  -> required by alias all
+  -> required by alias default
+  [1]
+