+++ test/blackbox-tests/test-cases/coq/compose-projects-cycle.t/A/a.v
@@ -1,0 +1,3 @@
+From C Require Import c.
+
+Inductive Hello := World | Bye.